Adjust Mappers configuration
- add ProfileMapper<TProfile>
- adjust Profiles and Mappers namespaces
