+++ src/Feature/Account/tests/Mappers/Profiles/AddressProfileTests.cs
@@ -12,19 +12,24 @@
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
 
-namespace Wooli.Feature.Account.Mappers
+namespace Wooli.Feature.Account.Tests.Mappers.Profiles
 {
+    using Account.Mappers.Profiles;
+
     using AutoMapper;
 
-    using Foundation.Commerce.Models.Entities.Addresses;
+    using Xunit;
 
-    using Models.Requests;
-
-    public class AddressProfile : Profile
+    public class AddressProfileTests
     {
-        public AddressProfile()
+        [Fact]
+        public void Configuration_ShouldBeValid()
         {
-            this.CreateMap<AddressRequest, Address>();
+            // arrange
+            var configuration = new MapperConfiguration(cfg => cfg.AddProfile<AddressProfile>());
+
+            // act, assert
+            configuration.AssertConfigurationIsValid();
         }
     }
 }
