+++ src/Feature/Account/tests/Wooli.Feature.Account.Tests.csproj
@@ -513,6 +513,7 @@
     <Compile Include="Controllers\AccountsControllerTests.cs" />
     <Compile Include="Controllers\AuthenticationControllerTests.cs" />
     <Compile Include="Mappers\AccountEntityMapperTests.cs" />
+    <Compile Include="Mappers\Profiles\AddressProfileTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
