+++ src/Feature/Catalog/tests/Wooli.Feature.Catalog.Tests.csproj
@@ -511,6 +511,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Controllers\SearchControllerTests.cs" />
+    <Compile Include="Mappers\Profiles\SearchProfileTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SimpleTest.cs" />
   </ItemGroup>
