+++ src/Foundation/Base/code/Wooli.Foundation.Base.csproj
@@ -460,6 +460,7 @@
     <Compile Include="Infrastructure\Pipelines\SafePipelineProcessor.cs" />
     <Compile Include="Mappers\IMapper.cs" />
     <Compile Include="Mappers\Mapper.cs" />
+    <Compile Include="Mappers\ProfileMapper.cs" />
     <Compile Include="Models\Configuration\Configuration.cs" />
     <Compile Include="Models\Logging\Log.cs" />
     <Compile Include="Models\Logging\CommonLog.cs" />
