+++ src/Foundation/Commerce/code/Mappers/Users/IUserMapper.cs
@@ -12,7 +12,7 @@
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
 
-namespace Wooli.Foundation.Commerce.ModelMappers.Users
+namespace Wooli.Foundation.Commerce.Mappers.Users
 {
     using Account.Infrastructure.Pipelines.Login;
