+++ src/Foundation/Commerce/code/Wooli.Foundation.Commerce.csproj
@@ -628,15 +628,15 @@
     <Compile Include="Infrastructure\Pipelines\SubmitVisitorOrder\SendEmailProcessor.cs" />
     <Compile Include="Mappers\Search\ISearchMapper.cs" />
     <Compile Include="Mappers\Search\SearchMapper.cs" />
-    <Compile Include="Mappers\Search\SearchProfile.cs" />
+    <Compile Include="Mappers\Profiles\SearchProfile.cs" />
     <Compile Include="ModelInitializers\OrderModelBuilder.cs" />
     <Compile Include="ModelInitializers\Contracts\IOrderModelBuilder.cs" />
     <Compile Include="ModelInitializers\Contracts\ICartModelBuilder.cs" />
     <Compile Include="ModelInitializers\CartModelBuilder.cs" />
     <Compile Include="ModelMappers\Contracts\IEntityMapper.cs" />
     <Compile Include="ModelMappers\EntityMapper.cs" />
-    <Compile Include="ModelMappers\Users\IUserMapper.cs" />
-    <Compile Include="ModelMappers\Users\UserMapper.cs" />
+    <Compile Include="Mappers\Users\IUserMapper.cs" />
+    <Compile Include="Mappers\Users\UserMapper.cs" />
     <Compile Include="Models.Generated.cs" />
     <Compile Include="Models\Account\ChangePasswordModel.cs" />
     <Compile Include="Models\Account\ChangePasswordResultModel.cs" />
@@ -698,7 +698,7 @@
     <Compile Include="Models\Entities\Search\SortOption.cs" />
     <Compile Include="Models\Entities\Cart\TotalPrice.cs" />
     <Compile Include="Models\Entities\Shipping\ShippingMethod.cs" />
-    <Compile Include="ModelMappers\Profiles\UserProfile.cs" />
+    <Compile Include="Mappers\Profiles\UserProfile.cs" />
     <Compile Include="Models\Region\CountryRegionModel.cs" />
     <Compile Include="Models\Region\SubdivisionModel.cs" />
     <Compile Include="Models\SortDirection.cs" />
