+++ src/Foundation/Connect/tests/Wooli.Foundation.Connect.Tests.csproj
@@ -562,7 +562,7 @@
     <Compile Include="Managers\Cart\CartManagerV2Tests.cs" />
     <Compile Include="Managers\Search\SearchManagerV2Tests.cs" />
     <Compile Include="Managers\Shipping\ShippingManagerV2Tests.cs" />
-    <Compile Include="Models\Mappings\SearchProfileTests.cs" />
+    <Compile Include="Mappers\Profiles\SearchProfileTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
